backend/requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.40.0
jinja2==3.1.4
python-dotenv==1.0.1
google-genai==1.56.0
reportlab==4.2.5
